\name{shift.sequence}
\alias{shift.sequence}
\title{shift.sequence}
\usage{
shift.sequence(v, places, dir="right")
}
\arguments{
\item{v}{Vector to shift}
\item{places}{The number of places to shift}
\item{dir}{Whether the shift should be right or left}
}
\description{
One of my functions to resolve issues for a similar function that seems to have been forgotten in haarfisz.
}
\details{
This function takes a sequence input and shifs it to the left or right by the specified number of places.
}
\value{
a shifted output sequence.
}
\author{Piotr Fryzlewicz}
\keyword{manip}
