\name{xquake}
\docType{data}
\alias{xquake}
\title{xquake}
\description{
Time series of the number of earthquakes of magnitude >= 3.0 which occurred in Northern California in 1024 weeks, the last week being 29/11 -- 05/12/2000. 
}
\usage{data(xquake)}
\source{
The series was composed using the data obtained from the Northern California Earthquake Data Center. 
}
\keyword{datasets}
