\name{hft.inv}
\alias{hft.inv}
\title{hft.inv}
\description{
Performs the inverse Haar-Fisz transform. 
}
\usage{
hft.inv(data) 
}
\arguments{
\item{data}{Vector of length \eqn{2^J} where J is an integer} 
}
\value{
Returns: 
\item{ihfx}{The inverse Haar-Fisz transform of \code{x} (vector of the same length as \code{data}).}  
}
\author{Piotr Fryzlewicz}
\keyword{manip}

